fix(precompiles): narrow migration skipped counter type
Amp-Thread-ID: https://ampcode.com/threads/T-019ccb1c-a4f7-72ef-a40f-f1d1b206a310

+++ crates/precompiles/src/validator_config_v2/mod.rs
@@ -73,7 +73,7 @@ pub struct ValidatorConfigV2 {
     address_to_index: Mapping<Address, u64>,
     pubkey_to_index: Mapping<B256, u64>,
     next_dkg_ceremony: u64,
-    migration_skipped_count: u64,
+    migration_skipped_count: u16,
     active_ingresses: Mapping<B256, bool>,
     active_indices: Vec<u64>,
 }
